+++ intellij_project/src/main/java/it/polimi/telco_webapp/services/OrdersAndCostsViewService.java
@@ -7,23 +7,25 @@ import jakarta.ejb.Stateless;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 
+import java.math.BigDecimal;
+
 @Stateless(name = "OrdersAndCostsView")
 public class OrdersAndCostsViewService {
     @PersistenceContext(unitName = "telco_webapp")
     private EntityManager em;
     public OrdersAndCostsViewService() {}
 
-    public float getBasePrice(int order_id) {
+    public BigDecimal getBasePrice(int order_id) {
         OrdersAndCostsView order = em.find(OrdersAndCostsView.class, order_id);
         return order.getBaseCost();
     }
 
-    public float getOptionsCost(int order_id) {
+    public BigDecimal getOptionsCost(int order_id) {
         OrdersAndCostsView order = em.find(OrdersAndCostsView.class, order_id);
         return order.getOptionsCost();
     }
 
-    public float getTotalCost(int order_id) {
+    public BigDecimal getTotalCost(int order_id) {
         OrdersAndCostsView order = em.find(OrdersAndCostsView.class, order_id);
         return order.getTotalCost();
     }
